# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Know Your Friend game API
 * OpenAPI spec version: 0.1.0
 */
export type CategorySuggestionStatus =
  (typeof CategorySuggestionStatus)[keyof typeof CategorySuggestionStatus];

export const CategorySuggestionStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export type LanguageCode = (typeof LanguageCode)[keyof typeof LanguageCode];

export const LanguageCode = {
  en: "en",
  de: "de",
  fr: "fr",
  es: "es",
  it: "it",
  ru: "ru",
} as const;

export interface CreateCategorySuggestionBody {
  language: LanguageCode;
  /**
   * @minLength 1
   * @maxLength 80
   */
  label: string;
  /**
   * @minLength 1
   * @maxLength 80
   */
  leftLabel: string;
  /**
   * @minLength 1
   * @maxLength 80
   */
  rightLabel: string;
}

export type UpdateCategorySuggestionStatusBodyStatus =
  (typeof UpdateCategorySuggestionStatusBodyStatus)[keyof typeof UpdateCategorySuggestionStatusBodyStatus];

export const UpdateCategorySuggestionStatusBodyStatus = {
  approved: "approved",
  rejected: "rejected",
} as const;

export interface UpdateCategorySuggestionStatusBody {
  status: UpdateCategorySuggestionStatusBodyStatus;
}

export type CategorySuggestion = CreateCategorySuggestionBody & {
  id: string;
  status: CategorySuggestionStatus;
  createdAt: string;
  updatedAt: string;
};

export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type Animal = (typeof Animal)[keyof typeof Animal];

export const Animal = {
  fox: "fox",
  raccoon: "raccoon",
  hedgehog: "hedgehog",
  deer: "deer",
  bear: "bear",
  wolf: "wolf",
  owl: "owl",
  squirrel: "squirrel",
} as const;

export interface CreateRoomBody {
  hostName: string;
  animal: Animal;
  totalRounds: number;
  language: LanguageCode;
}

export interface JoinRoomBody {
  playerName: string;
  animal: Animal;
}

export interface JoinRoomResponse {
  roomCode: string;
  playerId: string;
  playerToken: string;
  roomLanguage: LanguageCode;
}

export interface Player {
  id: string;
  name: string;
  animal?: Animal;
  score: number;
  isHost: boolean;
}

export type RoomStateStatus =
  (typeof RoomStateStatus)[keyof typeof RoomStateStatus];

export const RoomStateStatus = {
  waiting: "waiting",
  category_selection: "category_selection",
  self_rating: "self_rating",
  guessing: "guessing",
  round_results: "round_results",
  game_over: "game_over",
} as const;

export interface GuessResult {
  playerId: string;
  playerName: string;
  guess: number;
  selfRating: number;
  diff: number;
  /** Total round points, including bonusPoints. */
  points: number;
  /**
   * Perfect-guess bonus already included in points; zero for non-perfect guesses.
   * @minimum 0
   */
  bonusPoints?: number;
  /** Recorded slider extrema (direction-reversal points) from the guesser's drag, in chronological order. Always starts with the initial value and ends with the submitted guess. Empty if the guesser did not drag. */
  path?: number[];
  /** Same as `path`, but for the current player's self-rating drag. */
  selfRatingPath?: number[];
}

export interface Category {
  id: string;
  label: string;
  leftLabel: string;
  rightLabel: string;
}

export interface RoomState {
  roomCode: string;
  language: LanguageCode;
  status: RoomStateStatus;
  players: Player[];
  currentRound: number;
  totalRounds: number;
  roundsPerPlayer: number;
  /** @nullable */
  currentPlayerId: string | null;
  /** @nullable */
  currentCategory: string | null;
  /** @nullable */
  currentCategoryLabel: string | null;
  /** @nullable */
  currentCategoryLeftLabel: string | null;
  /** @nullable */
  currentCategoryRightLabel: string | null;
  /** @nullable */
  selfRating: number | null;
  guessesSubmitted: number;
  guessesTotal: number;
  /** @nullable */
  roundResults: GuessResult[] | null;
  availableCategories: Category[];
}

export type ListCategorySuggestionsParams = {
  status?: CategorySuggestionStatus;
};