# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Know Your Friend game API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: rooms
    description: Game room operations
  - name: category-suggestions
    description: Player-submitted category suggestions
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /rooms:
    post:
      operationId: createRoom
      tags: [rooms]
      summary: Create a new game room
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRoomBody"
      responses:
        "201":
          description: Room created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JoinRoomResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /rooms/{roomCode}/join:
    post:
      operationId: joinRoom
      tags: [rooms]
      summary: Join an existing game room
      parameters:
        - name: roomCode
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JoinRoomBody"
      responses:
        "200":
          description: Joined room
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JoinRoomResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Room not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /rooms/{roomCode}:
    get:
      operationId: getRoom
      tags: [rooms]
      summary: Get room state
      parameters:
        - name: roomCode
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Room state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomState"
        "404":
          description: Room not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /category-suggestions:
    post:
      operationId: createCategorySuggestion
      tags: [category-suggestions]
      summary: Submit a category suggestion for later review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategorySuggestionBody"
      responses:
        "201":
          description: Suggestion created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategorySuggestion"
        "400":
          description: Invalid suggestion
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    get:
      operationId: listCategorySuggestions
      tags: [category-suggestions]
      summary: List category suggestions for admin review
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/CategorySuggestionStatus"
      responses:
        "200":
          description: Suggestions ordered newest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategorySuggestion"

  /category-suggestions/{suggestionId}/status:
    patch:
      operationId: updateCategorySuggestionStatus
      tags: [category-suggestions]
      summary: Mark a category suggestion as approved or rejected
      parameters:
        - name: suggestionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCategorySuggestionStatusBody"
      responses:
        "200":
          description: Suggestion updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategorySuggestion"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Suggestion not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    CategorySuggestionStatus:
      type: string
      enum: [pending, approved, rejected]

    CreateCategorySuggestionBody:
      type: object
      additionalProperties: false
      properties:
        language:
          $ref: "#/components/schemas/LanguageCode"
        label:
          type: string
          minLength: 1
          maxLength: 80
        leftLabel:
          type: string
          minLength: 1
          maxLength: 80
        rightLabel:
          type: string
          minLength: 1
          maxLength: 80
      required: [language, label, leftLabel, rightLabel]

    UpdateCategorySuggestionStatusBody:
      type: object
      additionalProperties: false
      properties:
        status:
          type: string
          enum: [approved, rejected]
      required: [status]

    CategorySuggestion:
      allOf:
        - $ref: "#/components/schemas/CreateCategorySuggestionBody"
        - type: object
          properties:
            id:
              type: string
            status:
              $ref: "#/components/schemas/CategorySuggestionStatus"
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
          required: [id, status, createdAt, updatedAt]

    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    CreateRoomBody:
      type: object
      properties:
        hostName:
          type: string
        animal:
          $ref: "#/components/schemas/Animal"
        totalRounds:
          type: integer
        language:
          $ref: "#/components/schemas/LanguageCode"
      required:
        - hostName
        - animal
        - totalRounds
        - language

    JoinRoomBody:
      type: object
      properties:
        playerName:
          type: string
        animal:
          $ref: "#/components/schemas/Animal"
      required:
        - playerName
        - animal

    Animal:
      type: string
      enum:
        - fox
        - raccoon
        - hedgehog
        - deer
        - bear
        - wolf
        - owl
        - squirrel

    JoinRoomResponse:
      type: object
      properties:
        roomCode:
          type: string
        playerId:
          type: string
        playerToken:
          type: string
        roomLanguage:
          $ref: "#/components/schemas/LanguageCode"
      required:
        - roomCode
        - playerId
        - playerToken
        - roomLanguage

    LanguageCode:
      type: string
      enum:
        - en
        - de
        - fr
        - es
        - it
        - ru

    Player:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        animal:
          $ref: "#/components/schemas/Animal"
        score:
          type: integer
        isHost:
          type: boolean
      required:
        - id
        - name
        - score
        - isHost

    RoomState:
      type: object
      properties:
        roomCode:
          type: string
        language:
          $ref: "#/components/schemas/LanguageCode"
        status:
          type: string
          enum:
            - waiting
            - category_selection
            - self_rating
            - guessing
            - round_results
            - game_over
        players:
          type: array
          items:
            $ref: "#/components/schemas/Player"
        currentRound:
          type: integer
        totalRounds:
          type: integer
        roundsPerPlayer:
          type: integer
        currentPlayerId:
          type: ["string", "null"]
        currentCategory:
          type: ["string", "null"]
        currentCategoryLabel:
          type: ["string", "null"]
        currentCategoryLeftLabel:
          type: ["string", "null"]
        currentCategoryRightLabel:
          type: ["string", "null"]
        selfRating:
          type: ["integer", "null"]
        guessesSubmitted:
          type: integer
        guessesTotal:
          type: integer
        roundResults:
          type: ["array", "null"]
          items:
            $ref: "#/components/schemas/GuessResult"
        availableCategories:
          type: array
          items:
            $ref: "#/components/schemas/Category"
      required:
        - roomCode
        - language
        - status
        - players
        - currentRound
        - totalRounds
        - roundsPerPlayer
        - currentPlayerId
        - currentCategory
        - currentCategoryLabel
        - currentCategoryLeftLabel
        - currentCategoryRightLabel
        - selfRating
        - guessesSubmitted
        - guessesTotal
        - roundResults
        - availableCategories

    GuessResult:
      type: object
      properties:
        playerId:
          type: string
        playerName:
          type: string
        guess:
          type: integer
        selfRating:
          type: integer
        diff:
          type: integer
        points:
          type: integer
          description: Total round points, including bonusPoints.
        bonusPoints:
          type: integer
          minimum: 0
          description: Perfect-guess bonus already included in points; zero for non-perfect guesses.
        path:
          type: array
          description: Recorded slider extrema (direction-reversal points) from the guesser's drag, in chronological order. Always starts with the initial value and ends with the submitted guess. Empty if the guesser did not drag.
          items:
            type: integer
        selfRatingPath:
          type: array
          description: Same as `path`, but for the current player's self-rating drag.
          items:
            type: integer
      required:
        - playerId
        - playerName
        - guess
        - selfRating
        - diff
        - points

    Category:
      type: object
      properties:
        id:
          type: string
        label:
          type: string
        leftLabel:
          type: string
        rightLabel:
          type: string
      required:
        - id
        - label
        - leftLabel
        - rightLabel