# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Know Your Friend game API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Create a new game room
 */
export const CreateRoomBody = zod.object({
  hostName: zod.string(),
  animal: zod.enum([
    "fox",
    "raccoon",
    "hedgehog",
    "deer",
    "bear",
    "wolf",
    "owl",
    "squirrel",
  ]),
  totalRounds: zod.number(),
  language: zod.enum(["en", "de", "fr", "es", "it", "ru"]),
});

/**
 * @summary Join an existing game room
 */
export const JoinRoomParams = zod.object({
  roomCode: zod.coerce.string(),
});

export const JoinRoomBody = zod.object({
  playerName: zod.string(),
  animal: zod.enum([
    "fox",
    "raccoon",
    "hedgehog",
    "deer",
    "bear",
    "wolf",
    "owl",
    "squirrel",
  ]),
});

export const JoinRoomResponse = zod.object({
  roomCode: zod.string(),
  playerId: zod.string(),
  playerToken: zod.string(),
  roomLanguage: zod.enum(["en", "de", "fr", "es", "it", "ru"]),
});

/**
 * @summary Get room state
 */
export const GetRoomParams = zod.object({
  roomCode: zod.coerce.string(),
});

export const getRoomResponseRoundResultsItemBonusPointsMin = 0;

export const GetRoomResponse = zod.object({
  roomCode: zod.string(),
  language: zod.enum(["en", "de", "fr", "es", "it", "ru"]),
  status: zod.enum([
    "waiting",
    "category_selection",
    "self_rating",
    "guessing",
    "round_results",
    "game_over",
  ]),
  players: zod.array(
    zod.object({
      id: zod.string(),
      name: zod.string(),
      animal: zod
        .enum([
          "fox",
          "raccoon",
          "hedgehog",
          "deer",
          "bear",
          "wolf",
          "owl",
          "squirrel",
        ])
        .optional(),
      score: zod.number(),
      isHost: zod.boolean(),
    }),
  ),
  currentRound: zod.number(),
  totalRounds: zod.number(),
  roundsPerPlayer: zod.number(),
  currentPlayerId: zod.string().nullable(),
  currentCategory: zod.string().nullable(),
  currentCategoryLabel: zod.string().nullable(),
  currentCategoryLeftLabel: zod.string().nullable(),
  currentCategoryRightLabel: zod.string().nullable(),
  selfRating: zod.number().nullable(),
  guessesSubmitted: zod.number(),
  guessesTotal: zod.number(),
  roundResults: zod
    .array(
      zod.object({
        playerId: zod.string(),
        playerName: zod.string(),
        guess: zod.number(),
        selfRating: zod.number(),
        diff: zod.number(),
        points: zod
          .number()
          .describe("Total round points, including bonusPoints."),
        bonusPoints: zod
          .number()
          .min(getRoomResponseRoundResultsItemBonusPointsMin)
          .optional()
          .describe(
            "Perfect-guess bonus already included in points; zero for non-perfect guesses.",
          ),
        path: zod
          .array(zod.number())
          .optional()
          .describe(
            "Recorded slider extrema (direction-reversal points) from the guesser's drag, in chronological order. Always starts with the initial value and ends with the submitted guess. Empty if the guesser did not drag.",
          ),
        selfRatingPath: zod
          .array(zod.number())
          .optional()
          .describe(
            "Same as `path`, but for the current player's self-rating drag.",
          ),
      }),
    )
    .nullable(),
  availableCategories: zod.array(
    zod.object({
      id: zod.string(),
      label: zod.string(),
      leftLabel: zod.string(),
      rightLabel: zod.string(),
    }),
  ),
});

/**
 * @summary Submit a category suggestion for later review
 */
export const createCategorySuggestionBodyLabelMax = 80;

export const createCategorySuggestionBodyLeftLabelMax = 80;

export const createCategorySuggestionBodyRightLabelMax = 80;

export const CreateCategorySuggestionBody = zod.object({
  language: zod.enum(["en", "de", "fr", "es", "it", "ru"]),
  label: zod.string().min(1).max(createCategorySuggestionBodyLabelMax),
  leftLabel: zod.string().min(1).max(createCategorySuggestionBodyLeftLabelMax),
  rightLabel: zod
    .string()
    .min(1)
    .max(createCategorySuggestionBodyRightLabelMax),
});

/**
 * @summary List category suggestions for admin review
 */
export const ListCategorySuggestionsQueryParams = zod.object({
  status: zod.enum(["pending", "approved", "rejected"]).optional(),
});

export const listCategorySuggestionsResponseOneLabelMax = 80;

export const listCategorySuggestionsResponseOneLeftLabelMax = 80;

export const listCategorySuggestionsResponseOneRightLabelMax = 80;

export const ListCategorySuggestionsResponseItem = zod
  .object({
    language: zod.enum(["en", "de", "fr", "es", "it", "ru"]),
    label: zod.string().min(1).max(listCategorySuggestionsResponseOneLabelMax),
    leftLabel: zod
      .string()
      .min(1)
      .max(listCategorySuggestionsResponseOneLeftLabelMax),
    rightLabel: zod
      .string()
      .min(1)
      .max(listCategorySuggestionsResponseOneRightLabelMax),
  })
  .and(
    zod.object({
      id: zod.string(),
      status: zod.enum(["pending", "approved", "rejected"]),
      createdAt: zod.coerce.date(),
      updatedAt: zod.coerce.date(),
    }),
  );
export const ListCategorySuggestionsResponse = zod.array(
  ListCategorySuggestionsResponseItem,
);

/**
 * @summary Mark a category suggestion as approved or rejected
 */
export const UpdateCategorySuggestionStatusParams = zod.object({
  suggestionId: zod.coerce.string(),
});

export const UpdateCategorySuggestionStatusBody = zod.object({
  status: zod.enum(["approved", "rejected"]),
});

export const updateCategorySuggestionStatusResponseOneLabelMax = 80;

export const updateCategorySuggestionStatusResponseOneLeftLabelMax = 80;

export const updateCategorySuggestionStatusResponseOneRightLabelMax = 80;

export const UpdateCategorySuggestionStatusResponse = zod
  .object({
    language: zod.enum(["en", "de", "fr", "es", "it", "ru"]),
    label: zod
      .string()
      .min(1)
      .max(updateCategorySuggestionStatusResponseOneLabelMax),
    leftLabel: zod
      .string()
      .min(1)
      .max(updateCategorySuggestionStatusResponseOneLeftLabelMax),
    rightLabel: zod
      .string()
      .min(1)
      .max(updateCategorySuggestionStatusResponseOneRightLabelMax),
  })
  .and(
    zod.object({
      id: zod.string(),
      status: zod.enum(["pending", "approved", "rejected"]),
      createdAt: zod.coerce.date(),
      updatedAt: zod.coerce.date(),
    }),
  );