# lib/api-zod/src/generated/types/categorySuggestion.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Know Your Friend game API
 * OpenAPI spec version: 0.1.0
 */
import type { CategorySuggestionStatus } from "./categorySuggestionStatus";
import type { CreateCategorySuggestionBody } from "./createCategorySuggestionBody";

export type CategorySuggestion = CreateCategorySuggestionBody & {
  id: string;
  status: CategorySuggestionStatus;
  createdAt: Date;
  updatedAt: Date;
};