# lib/api-zod/src/generated/types/categorySuggestionStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Know Your Friend game API
 * OpenAPI spec version: 0.1.0
 */

export type CategorySuggestionStatus =
  (typeof CategorySuggestionStatus)[keyof typeof CategorySuggestionStatus];

export const CategorySuggestionStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;