# lib/api-zod/src/generated/types/createCategorySuggestionBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Know Your Friend game API
 * OpenAPI spec version: 0.1.0
 */
import type { LanguageCode } from "./languageCode";

export interface CreateCategorySuggestionBody {
  language: LanguageCode;
  /**
   * @minLength 1
   * @maxLength 80
   */
  label: string;
  /**
   * @minLength 1
   * @maxLength 80
   */
  leftLabel: string;
  /**
   * @minLength 1
   * @maxLength 80
   */
  rightLabel: string;
}