# lib/api-zod/src/generated/types/updateCategorySuggestionStatusBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Know Your Friend game API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateCategorySuggestionStatusBodyStatus } from "./updateCategorySuggestionStatusBodyStatus";

export interface UpdateCategorySuggestionStatusBody {
  status: UpdateCategorySuggestionStatusBodyStatus;
}